Re-org 2014 tests
