+++ tests/test_species/test_human.py
@@ -1,12 +1,11 @@
 import unittest
 
-from charsheets.character import Character, Character2014
-from charsheets.constants import Skill, Feature, Language, Stat
+from charsheets.character import Character
+from charsheets.constants import Skill, Feature, Language
 from charsheets.features import Alert
 from charsheets.main import render
 from charsheets.species import Human, Skillful, Versatile
-from charsheets.race2014 import Human as Human14
-from tests.dummy import DummyCharClass, DummyOrigin, DummyBackground
+from tests.dummy import DummyOrigin
 
 
 #######################################################################
@@ -49,33 +48,6 @@ class TestHuman(unittest.TestCase):
     ###################################################################
     def test_versatile(self):
         self.assertTrue(self.c.has_feature(Feature.ALERT))
-
-
-#######################################################################
-class TestHuman14(unittest.TestCase):
-    def setUp(self):
-        self.c = Character2014(
-            "test_human",
-            DummyBackground(),
-            Human14(Language.ORC),
-            strength=10,
-            dexterity=10,
-            constitution=10,
-            intelligence=10,
-            wisdom=10,
-            charisma=10,
-        )
-
-    ###################################################################
-    def test_stats(self):
-        self.assertEqual(self.c.speed.value, 30)
-        self.assertEqual(int(self.c.stats[Stat.STRENGTH].value), 11)
-        self.assertEqual(int(self.c.stats[Stat.DEXTERITY].value), 11)
-        self.assertEqual(int(self.c.stats[Stat.CONSTITUTION].value), 11)
-        self.assertEqual(int(self.c.stats[Stat.INTELLIGENCE].value), 11)
-        self.assertEqual(int(self.c.stats[Stat.WISDOM].value), 11)
-        self.assertEqual(int(self.c.stats[Stat.CHARISMA].value), 11)
-        self.assertIn(Language.ORC, self.c.languages)
 
 
 #######################################################################
