+++ tests/tests2014/test_classes2014/test_paladin.py
